web-browser-sq2018-echooo/requirements.txt
requests==2.18.4
Pillow==5.0.0
